pyproject.toml
[project]
name = "mushrooms-e1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.9.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "executing==2.2.0",
    "fastapi>=0.103.1",
    "fastjsonschema==2.21.1",
    "fqdn==1.5.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "json5==0.12.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter==1.1.1",
    "jupyter-client==8.6.3",
    "jupyter-console==6.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.15.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.2",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.15",
    "markupsafe==3.0.2",
    "matplotlib>=3.10.1",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.4.2",
    "notebook-shim==0.2.4",
    "numpy>=1.25.2",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas>=2.1.0",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "platformdirs==4.3.7",
    "prometheus-client==0.21.1",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "psycopg2-binary>=2.9.10",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydantic>=2.3.0",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-json-logger==3.3.0",
    "python-multipart>=0.0.6",
    "pytz==2025.2",
    "pywin32==310",
    "pywinpty==2.0.15",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rpds-py==0.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "send2trash==1.8.3",
    "setuptools==80.3.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy>=2.0.40",
    "stack-data==0.6.3",
    "terminado==0.18.1",
    "tinycss2==1.4.0",
    "tornado==6.4.2",
    "traitlets==5.14.3",
    "types-python-dateutil==2.9.0.20241206",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.4.0",
    "uvicorn>=0.23.2",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "widgetsnbextension==4.0.14",
    "xgboost>=1.7.6",
]
